Revert "Move messages to a separate file"
This reverts commit 36f61fe81899d503538e81e1b9dce8f9304d74fb.
